requirements.txt
numpy
pandas
python-dotenv
pymongo
certifi
scikit-learn

# -e .